requirements.txt
Django==3.0.7
django-filter==2.3.0
djangorestframework==3.11.1
jsonpath==0.82
Markdown==3.2.2
pytest==5.4.1
pytz==2018.5
requests==2.24.0
sqlparse==0.3.1